# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    post:
      operationId: ArtistsController_create
      parameters: []
      requestBody:
        required: true
        description: A name and grammy for the new artist
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "201":
          description: The artist has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Request body does not contain required fields
      tags: &a2
        - artist
    get:
      operationId: ArtistsController_findAll
      parameters: []
      responses:
        "200":
          description: A artists has been successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
      tags: *a2
  "/artist/{id}":
    get:
      operationId: ArtistsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a artist that exists in the database
          schema:
            type: string
      responses:
        "200":
          description: A artist has been successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: A artist with given id is invalid (not uuid).
        "404":
          description: A artist with given id does not exist.
      tags: *a2
    put:
      operationId: ArtistsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a artist that exists in the database
          schema:
            type: string
      requestBody:
        required: true
        description: A new name and grammy for the artist
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
      responses:
        "200":
          description: A artist has been successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: A artist with given id is invalid (not uuid).
        "404":
          description: A artist with given id does not exist.
      tags: *a2
    delete:
      operationId: ArtistsController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a artist that exists in the database
          schema:
            type: string
      responses:
        "204":
          description: A artist has been successfully deleted
        "400":
          description: A artist with given id is invalid (not uuid).
        "404":
          description: A artist with given id does not exist.
      tags: *a2
...
  /album:
    post:
      operationId: AlbumsController_create
      parameters: []
      requestBody:
        required: true
        description: A name, year and artistId for the new album
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "201":
          description: The album has been successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Request body does not contain required fields
      tags: &a4
        - album
    get:
      operationId: AlbumsController_findAll
      parameters: []
      responses:
        "200":
          description: A albums has been successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
      tags: *a4
  "/album/{id}":
    get:
      operationId: AlbumsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a album that exists in the database
          schema:
            type: string
      responses:
        "200":
          description: A album has been successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: A album with given id is invalid (not uuid).
        "404":
          description: A album with given id does not exist.
      tags: *a4
    put:
      operationId: AlbumsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a album that exists in the database
          schema:
            type: string
      requestBody:
        required: true
        description: A new name, year or artistId for the album
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: A album has been successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: A album with given id is invalid (not uuid).
        "404":
          description: A album with given id does not exist.
      tags: *a4
    delete:
      operationId: AlbumsController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a album that exists in the database
          schema:
            type: string
      responses:
        "204":
          description: A album has been successfully deleted
        "400":
          description: A album with given id is invalid (not uuid).
        "404":
          description: A album with given id does not exist.
      tags: *a4
...
  "/favs/album/{id}":
    post:
      operationId: AlbumToFavoritesController_createAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a album that exists in the database
          schema:
            type: string
      responses:
        "201":
          description: A album has been successfully added
        "400":
          description: A album with given id is invalid (not uuid).
        "422":
          description: A album with given id does not exist
      tags: &a5
        - favs/album
    delete:
      operationId: AlbumToFavoritesController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a album that exists in the database
          schema:
            type: string
      responses:
        "204":
          description: A album has been successfully deleted
        "400":
          description: A album with given id is invalid (not uuid).
        "404":
          description: A album with given id does not exist.
      tags: *a5
...
  "/favs/artist/{id}":
    post:
      operationId: ArtistToFavoritesController_createArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a artist that exists in the database
          schema:
            type: string
      responses:
        "201":
          description: A artist has been successfully added
        "400":
          description: A artist with given id is invalid (not uuid).
        "422":
          description: A artist with given id does not exist
      tags: &a7
        - favs/artist
    delete:
      operationId: ArtistToFavoritesController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Should be an id of a artist that exists in the database
          schema:
            type: string
      responses:
        "204":
          description: A artist has been successfully deleted
        "400":
          description: A artist with given id is invalid (not uuid).
        "404":
          description: A artist with given id does not exist.
      tags: *a7
info:
  title: Home Library Service
  description: Home music library service OpenAPI 3.0
  version: "1.0"
  contact: {}
...
servers:
  - url: http://localhost:4000
  - url: http://localhost
components:
  schemas:
...
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        grammy:
          type: boolean
      required:
        - id
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
...
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
      required:
        - name
        - year
    Album:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
      required:
        - id
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
...